Restructuring for pip-packaging: - Added setup configuration. - Renamed `code` to `faultanalysistoolbox` for package naming. - Removed pycache files.
