+++ setup.cfg
@@ -1,0 +1,21 @@
+[metadata]
+name = faultanalysistoolbox
+version = 0.0.1
+
+[options]
+packages=find:
+install_requires=
+    numpy
+    scipy
+    pandas
+    matplotlib
+    meshio
+    network
+    opencv-python
+    Pillow
+    seaborn
+    vtk
+    cv_algorithms@git+https://github.com/ulikoehler/cv_algorithms.git
+
+[options.packages.find]
+exclude=examples
